DDR4 command in I2C mode

+++ src/pirate/hwi2c_pio.c
@@ -389,7 +389,7 @@ bool i2c_write(uint8_t addr, uint8_t *data, uint16_t len) {
     return false;
 }
 
-bool i2c_read(uint8_t addr, uint8_t *data, uint8_t len) {
+bool i2c_read(uint8_t addr, uint8_t *data, uint16_t len) {
     hwi2c_status_t i2c_result = pio_i2c_read_array_timeout(addr | 1u, data, len, 0xfffffu);
     if(i2c_result != HWI2C_OK) {
         if(i2c_result == HWI2C_TIMEOUT) {
